+++ config/config.go
@@ -31,10 +31,11 @@ type CertSource struct {
 
 type Listen struct {
 	Addr         string
-	Scheme       string
+	Proto        string
 	ReadTimeout  time.Duration
 	WriteTimeout time.Duration
 	CertSource   CertSource
+	StrictMatch  bool
 }
 
 type UI struct {
@@ -67,11 +68,17 @@ type Runtime struct {
 }
 
 type Metrics struct {
-	Target       string
-	Prefix       string
-	Interval     time.Duration
-	GraphiteAddr string
-	StatsDAddr   string
+	Target           string
+	Prefix           string
+	Names            string
+	Interval         time.Duration
+	GraphiteAddr     string
+	StatsDAddr       string
+	CirconusAPIKey   string
+	CirconusAPIApp   string
+	CirconusAPIURL   string
+	CirconusCheckID  string
+	CirconusBrokerID string
 }
 
 type Registry struct {