+++ metrics/metrics_test.go
@@ -6,11 +6,25 @@ import (
 	"testing"
 )
 
-func TestDefaultPrefix(t *testing.T) {
+func TestParsePrefix(t *testing.T) {
 	hostname = func() (string, error) { return "myhost", nil }
 	os.Args = []string{"./myapp"}
-	if got, want := defaultPrefix(), "myhost.myapp"; got != want {
-		t.Errorf("got %v want %v", got, want)
+	got, err := parsePrefix("{{clean .Hostname}}.{{clean .Exec}}")
+	if err != nil {
+		t.Fatalf("%v", err)
+	}
+	want := "myhost.myapp"
+	if got != want {
+		t.Errorf("ParsePrefix: got %v want %v", got, want)
+	}
+
+	got, err = parsePrefix("default")
+	if err != nil {
+		t.Fatalf("%v", err)
+	}
+	want = "myhost.myapp"
+	if got != want {
+		t.Errorf("ParsePrefix Old default style: got %v want %v", got, want)
 	}
 }
 
@@ -32,7 +46,12 @@ func TestTargetName(t *testing.T) {
 		if err != nil {
 			t.Fatalf("%d: %v", i, err)
 		}
-		if got, want := TargetName(tt.service, tt.host, tt.path, u), tt.name; got != want {
+
+		got, err := TargetName(tt.service, tt.host, tt.path, u)
+		if err != nil {
+			t.Fatalf("%d: %v", i, err)
+		}
+		if want := tt.name; got != want {
 			t.Errorf("%d: got %q want %q", i, got, want)
 		}
 	}